fixed file structure and also , refactored some functions
