+++ mobile/Yohannes Tigistu/ecommerce_app/lib/core/usecases/usecase.dart
@@ -1,0 +1,7 @@
+import 'package:dartz/dartz.dart';
+
+import '../errors/failures.dart';
+
+abstract class UseCase<Type, Params> {
+  Future<Either<Failures, Type>> call(Params params);
+}
