+++ mobile/Yohannes Tigistu/ecommerce_app/lib/features/Products/domain/usecases/update_product_usecase.dart
@@ -1,16 +1,28 @@
 import 'package:dartz/dartz.dart';
+import 'package:equatable/equatable.dart';
+import 'package:flutter/widgets.dart';
 
 import '../../../../core/errors/failures.dart';
 import '../../../../core/success/confirmation.dart';
+import '../../../../core/usecases/usecase.dart';
 import '../entities/product.dart';
 import '../repositories/product_repository.dart';
 
-class UpdateProductUsecase {
+class UpdateProductUsecase implements UseCase<Confirmation, Params>{
   final ProductRepository repository;
 
   UpdateProductUsecase(this.repository);
-
-  Future<Either<Failures, Confirmation>> call(Product product) async {
-    return await repository.updateProduct(product);
+ @override
+  Future<Either<Failures, Confirmation>> call(Params params) async {
+    return await repository.updateProduct(params.product);
   }
 } 
+class Params extends Equatable {
+  final Product product;
+
+  Params({required this.product});
+
+  @override
+  List<Object?> get props => [product];
+}
+
