+++ mobile/Yohannes Tigistu/ecommerce_app/lib/features/Products/domain/usecases/view_specific_product_usecase.dart
@@ -1,14 +1,27 @@
 import 'package:dartz/dartz.dart';
+import 'package:equatable/equatable.dart';
 
 import '../../../../core/errors/failures.dart';
+import '../../../../core/usecases/usecase.dart';
 import '../entities/product.dart';
 import '../repositories/product_repository.dart';
 
-class ViewSpecificProductUsecase {
+
+class ViewSpecificProductUsecase implements UseCase <Product , Params> {
   final ProductRepository repository;
 
   ViewSpecificProductUsecase(this.repository);
-  Future<Either<Failures, Product>> call(int id) async{
-    return repository.getProductById(id);
+  @override
+  Future<Either<Failures, Product>> call(Params params ) async{
+    return repository.getProductById(params.id);
   }
+}
+class Params extends Equatable{
+ final int id;
+
+ const Params({required this.id});
+
+ @override
+ List<Object> get props => [id];
+
 }
