+++ mobile/Yohannes Tigistu/ecommerce_app/test/features/product/data/models/product_model_test.dart
@@ -5,7 +5,7 @@ import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../../features/Products/data/models/product_model.dart';
+import '../../../../../lib/features/Products/data/models/product_model.dart';
 import '../../../../fixtures/fixture_reader.dart';
 
 void main(){
